// src/Interfaces/contact.ts
interface Contact { 
    id?: string;
    firstName: string;
    lastName: string;
    favorite: boolean;
    picture?: string;
    phoneNumber?: string;
    description?: string;
    facebook?: string;
    twitter?: string;
    instagram?: string;
    github?: string;
    linkedin?: string;
}

export default Contact;